Handle exceptions threw by enqueuers (#28)
