+++ src/EventSubscriber/ChannelSubscriber.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace Webgriffe\SyliusActiveCampaignPlugin\EventSubscriber;
 
+use Psr\Log\LoggerInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
 use Symfony\Component\Messenger\MessageBusInterface;
+use Throwable;
 use Webgriffe\SyliusActiveCampaignPlugin\Enqueuer\ConnectionEnqueuerInterface;
 use Webgriffe\SyliusActiveCampaignPlugin\Message\Connection\ConnectionRemove;
 use Webgriffe\SyliusActiveCampaignPlugin\Model\ActiveCampaignAwareInterface;
@@ -16,7 +18,8 @@ final class ChannelSubscriber implements EventSubscriberInterface
 {
     public function __construct(
         private MessageBusInterface $messageBus,
-        private ConnectionEnqueuerInterface $connectionEnqueuer
+        private ConnectionEnqueuerInterface $connectionEnqueuer,
+        private LoggerInterface $logger
     ) {
     }
 
@@ -35,7 +38,12 @@ final class ChannelSubscriber implements EventSubscriberInterface
         if (!$channel instanceof ChannelInterface || !$channel instanceof ActiveCampaignAwareInterface) {
             return;
         }
-        $this->connectionEnqueuer->enqueue($channel);
+
+        try {
+            $this->connectionEnqueuer->enqueue($channel);
+        } catch (Throwable $throwable) {
+            $this->logger->error($throwable->getMessage(), $throwable->getTrace());
+        }
     }
 
     public function removeChannel(GenericEvent $event): void
