+++ src/Resources/config/services.xml
@@ -9,6 +9,7 @@
         <import resource="services/factory.xml"/>
         <import resource="services/generator.xml"/>
         <import resource="services/http_client.xml"/>
+        <import resource="services/logger.xml"/>
         <import resource="services/mapper.xml"/>
         <import resource="services/message_handler.xml"/>
         <import resource="services/middleware.xml"/>
