+++ src/Resources/config/services/logger.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <parameters>
+        <parameter key="webgriffe_sylius_active_campaign.logger.channel_name">webgriffe_active_campaign_plugin</parameter>
+    </parameters>
+    <services>
+        <service id="webgriffe.sylius_active_campaign_plugin.logger" class="Monolog\Logger">
+            <argument type="string">%webgriffe_sylius_active_campaign.logger.channel_name%</argument>
+        </service>
+    </services>
+</container>
