--- tests/DataFixtures/ORM/resources/EventSubscriber/CheckoutSubscriberTest/channels.yaml
@@ -1,17 +1,0 @@
-App\Entity\Channel\Channel:
-    fashion_shop:
-        code: "fashion_shop"
-        name: "Fashion shop"
-        taxCalculationStrategy: "order_items_based"
-        defaultLocale: "@en_US"
-        baseCurrency: "@EUR"
-
-Sylius\Component\Locale\Model\Locale:
-    en_US:
-        code: "en_US"
-    it_IT:
-        code: "it_IT"
-
-Sylius\Component\Currency\Model\Currency:
-    EUR:
-        code: "EUR"
