--- tests/DataFixtures/ORM/resources/EventSubscriber/CheckoutSubscriberTest/customers.yaml
@@ -1,5 +1,0 @@
-App\Entity\Customer\Customer:
-    customer_01:
-        email: 'jim@email.com'
-        first_name: 'James'
-        last_name: 'Rodriguez'
